chore: extract components folder
