--- src/app/components/Dashboard.js
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import App from './Dashboard';
-
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
