+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import App from './components/Dashboard';
+import App from '../components/App';
 
 export default function Page() {
     return <App />
